-Reduced duplicated code
