+++ src/views/blockapps-intro/GetBalanceView.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import AceEditor from 'react-ace'
 import Hightlight from 'react-highlight'
-import Ace from 'brace'
 import BaseView from './BaseView'
 
 import 'brace/mode/html'
@@ -9,9 +8,6 @@ import 'brace/mode/javascript'
 import 'brace/theme/github'
 
 import 'highlight.js/styles/default.css'
-
-const Anchor = Ace.acequire('ace/anchor').Anchor
-const Range = Ace.acequire('ace/range').Range
 
 export class GetBalanceView extends BaseView {
     static defaultProps = {
@@ -44,72 +40,12 @@ export class GetBalanceView extends BaseView {
 
     constructor (props) {
       super(props)
-      this.state = {
-        editor: {},
-        editableRanges: [],
-        userCode: this.props.template
-      }
-      this.handleProceed = this.handleProceed.bind(this)
-      this.handleKeyboard = this.handleKeyboard.bind(this)
-      this.onEditorLoad = this.onEditorLoad.bind(this)
-      this.onEditorChange = this.onEditorChange.bind(this)
-    }
-
-    static state = {
-      userCode: 'usercode'
-    }
-
-    static propTypes = {
-      step: React.PropTypes.number,
-      history: React.PropTypes.object,
-      codeBlock: React.PropTypes.string,
-      editableArea: React.PropTypes.array,
-      template: React.PropTypes.string
-    }
-
-    onEditorChange (newValue) {
-      this.state.userCode = newValue
-    }
-
-    onEditorLoad (editor) {
-      this.state.editor = editor
-      this.state.doc = editor.getSession().getDocument()
-
-      editor.renderer.setScrollMargin(10, 10, 10, 10)
-      editor.getSession().setUseWorker(false)
-      editor.moveCursorTo(11, 0)
-
-      let startAnchor, endAnchor, editableRange
-      for (let area of this.props.editableArea) {
-        startAnchor = new Anchor(this.state.doc, area[0], area[1])
-        endAnchor = new Anchor(this.state.doc, area[2], area[3])
-        editableRange = Range.fromPoints(startAnchor.getPosition(), endAnchor.getPosition())
-        editor.getSession().addMarker(editableRange, 'ace_active-line', 'fullline')
-        this.state.editableRanges.push(editableRange)
-      }
-
-      editor.keyBinding.addKeyboardHandler({
-        handleKeyboard: this.handleKeyboard
-      })
-    }
-
-    handleKeyboard (data, hash, keyString, keyCode, event) {
-      if (hash === -1 || (keyCode <= 40 && keyCode >= 37)) return false
-
-      const cursorPosition = this.state.editor.getCursorPosition()
-      let insideReadOnly = false
-      for (let editableRange of this.state.editableRanges) {
-        insideReadOnly = !editableRange.contains(cursorPosition.row, cursorPosition.column)
-      }
-      if (insideReadOnly) {
-        return {command: 'null', passEvent: false}
-      }
     }
 
     handleProceed () {
       let userInput = this.state.userCode.split('\n')[11].trim()
       if (userInput === this.props.codeBlock) {
-        this.props.history.push('/tutorial/step/3')
+        this.props.history.push('/tutorial/step/' + (this.props.step + 1))
       } else {
         this.state.invalidInput = true
       }
@@ -140,7 +76,7 @@ export class GetBalanceView extends BaseView {
                     highlightActiveLine={false}
                     width='85%'
                     height='30em'
-                    name='includeLibraryEditor'
+                    name='tutorialEditor'
                     value={this.props.template}
                 />
             </div>
